More work on the new Batches API for AG
